requirements.txt
django==1.6.5
psycopg2==2.5.3
South==0.8.4
six==1.7.2
django-dbsettings==0.6
django-phonenumber-field==0.5
django-celery==3.1.10
django-celery-with-redis==3.0